Fix bad request (error 500) temporarily
The problem with most api requests returning error 500 is that with a
production server, which uses multiple threads and multiple processes,
such as uwsgi, sharing of global variables is neither thread or process
safe.

Just limit the number of processes to 1 and disable threads for now so
that the site at least loads. Promise I'll get back to it later and do a
proper fix (using redis or memcached are probably good options).